[Refactor] 클라이언트 렌더링 코드 사용 수정
- page.tsx에서 'use client' 키워드 삭제
- 접근성 향상을 위해 Logo 대체 텍스트 수정
- layout.tsx에 Head 및 title 추가

Resolves #105
