+++ src/app/signup/page.tsx
@@ -1,20 +1,7 @@
-'use client';
-
-import SignUpKind from '@/features/sign/components/signup-kind';
-import { motion } from 'framer-motion';
+import SignUp from '@/features/sign/components/sign-up';
 
 function SignUpPage() {
-  return (
-    <div className="text-white">
-      <motion.div
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: -20 }}
-        transition={{ duration: 1 }}
-      >
-        <SignUpKind />
-      </motion.div>
-    </div>
-  );
+  return <SignUp />;
 }
 
 export default SignUpPage;
